% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognizer.R
\docType{package}
\name{cognizer}
\alias{cognizer}
\alias{cognizer-package}
\title{R Wrapper for IBM Watson Services}
\description{
cognizeR provides interface to IBM Watson services that can
  process text, image and audio data.
}

