% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cognizers.R
\name{text_audio}
\alias{text_audio}
\title{IBM Text-to-speech API.}
\usage{
text_audio(text, userpwd, keep_data = "true", directory,
  voice = "en-US_AllisonVoice", accept = "audio/ogg;codecs=opus")
}
\arguments{
\item{text}{Character vector containing strings to be processed.}

\item{userpwd}{Character scalar that contains 'username:password' string.}

\item{keep_data}{Character scalar specifying whether to share your data with
Watson services for the purpose of training their models.}

\item{directory}{Character scalar specifying directory for storing audio files.}

\item{voice}{Character scalar setting language and voice model for the synthesized
voice. Many models are available: de-DE_BirgitVoice, de-DE_DieterVoice,
en-GB_KateVoice, en-US_LisaVoice, en-US_MichaelVoice, es-ES_EnriqueVoice,
es-ES_LauraVoice, es-US_SofiaVoice, fr-FR_ReneeVoice, it-IT_FrancescaVoice,
ja-JP_EmiVoice, pt-BR_IsabelaVoice.}

\item{accept}{Characte scalar specifying format for the audio. Alternatives are
audio/wav ,audio/flac, audio/l16, audio/basic.}
}
\value{
Audio file with selected format is saved into selected directory. The name
  is based on integer representation of UTF time and a number of characters of the
  processed text.

Logical scalar is returned invisibly.
}
\description{
Synthesizes an audio record from text.
}
\seealso{
Check \url{http://www.ibm.com/watson/developercloud/text-to-speech.html}
  for further documentation, and \url{https://text-to-speech-demo.mybluemix.net/}
  for a web demo.
}

