% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cognizers.R
\name{text_personality}
\alias{text_personality}
\title{IBM personality analysis of text}
\usage{
text_personality(text, userpwd, keep_data = "true", callback = NULL,
  model_id = "es-en-conversational", raw_scores = "false",
  consumption_preferences = "false", csv_headers = "false",
  version = "2020-01-01", content_type = "text/plain; charset=utf-8",
  content_language = "en", accept = "application/json",
  accept_language = "en")
}
\arguments{
\item{text}{Character vector.}

\item{userpwd}{Character scalar that contains 'username:password' string.}

\item{keep_data}{Character scalar specifying whether to share your data with
Watson services for the purpose of training their models.}

\item{callback}{Function that can be applied to responses to examine http status,
headers, and content, to debug or to write a custom parser for content.
The default callback parses content into a data.frame while dropping other
response values to make the output easily passable to tidyverse packages like
dplyr or ggplot2. For further details or debugging one can pass a fail or a
more compicated function.}

\item{model_id}{Character scalar formated as 'source-target-domain'.
Source language (Arabic, Brazilian Portuguese, English, French, Italian,
or Spanish), target language (Arabic, Brazilian Portuguese, English, French,
Italian, or Spanish) and domain of text (conversational, news, patent).
Check IBM documentation for other language mappings.}

\item{raw_scores}{Character scalar showing whether to include only normalized
statistics or also raw statistics.}

\item{consumption_preferences}{Character scalar showing whether to 
return consumption preferences}

\item{csv_headers}{Character scalar showing whether to return column labels when
Accept-Content is set to 'text/csv'.}

\item{version}{Character scalar giving date that specifies the algorithm that went
operational on or before the date. Future dates select the most recent algorithm.}

\item{content_type}{Character scalar setting input data type header. Alternatives
are 'application/json; charset=utf-8' and 'text/html; charset=ISO-8859-1'.}

\item{content_language}{Character scalar setting input language. Alternatives are
'ar' (Arabic), 'es' (Spanish), 'ja' (Japanese).}

\item{accept}{Character scalar that specifies response format. Alternative is
text/plain.}

\item{accept_language}{Character scalar setting output langauge. Alternatives are
'ar' (Arabic), 'de' (German), 'es' (Spanish), 'fr' (French), 'it' (Italian),
'ja' (Japanese), 'ko' (Korean), 'pt-br' (Brazilian Portuguese),
'zh-cn' (Simplified Chinese), 'zh-tw' (Traditional Chinese).}
}
\value{
List containing parsed content.
}
\description{
Analyze your text along the Big 5 dimensions of personality.
}
\seealso{
Check \url{http://www.ibm.com/watson/developercloud/doc/personality-insights/}
  for further documentation, and \url{https://personality-insights-livedemo.mybluemix.net/}
  for a web demo.
}

