% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cognizers.R
\name{text_tone}
\alias{text_tone}
\title{IBM Tone Analyzer of Text}
\usage{
text_tone(text, userpwd, keep_data = "true", callback = NULL,
  content_type = "text/plain; charset=utf-8", version = "2016-05-19",
  tones = "", sentences = "true")
}
\arguments{
\item{text}{Character vector.}

\item{userpwd}{Character scalar that contains 'username:password' string.}

\item{keep_data}{Character scalar specifying whether to share your data with
Watson services for the purpose of training their models.}

\item{callback}{Function that can be applied to responses to examine http status,
headers, and content, to debug or to write a custom parser for content.
The default callback parses content into a data.frame while dropping other
response values to make the output easily passable to tidyverse packages like
dplyr or ggplot2. For further details or debugging one can pass a print or a
more compicated function.}

\item{content_type}{Characte scalar specifying the HTTP header with type of text
and its encoding.}

\item{version}{Character scalar that specifies the data of most recent version of
the algorithm.}

\item{tones}{Character scalar that allows selecting one the three possible tones:
emotion, language, social.}

\item{sentences}{Character scalar specifying whether to do analysis at the
sentence level.}
}
\value{
Data.frame containing parsed content in a tidy fashion.
}
\description{
Infers three types of tone - emotion, language, social - from the
  whole text or at sentense level.
}
\seealso{
Check \url{http://www.ibm.com/watson/developercloud/doc/tone-analyzer/}
  for further documentation, and \url{https://tone-analyzer-demo.mybluemix.net/?cm_mc_uid=70865809903714586773519&cm_mc_sid_50200000=1468424667}
  for a web demo.
}

