% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cognizers.R
\name{text_translate}
\alias{text_translate}
\title{IBM Language Translation API.}
\usage{
text_translate(text, userpwd, keep_data = "true", callback = NULL,
  model_id = "es-en-conversational", accept = "application/json")
}
\arguments{
\item{text}{Character vector.}

\item{userpwd}{Character scalar that contains 'username:password' string.}

\item{keep_data}{Character scalar specifying whether to share your data with
Watson services for the purpose of training their models.}

\item{callback}{Function that can be applied to responses to examine http status,
headers, and content, to debug or to write a custom parser for content.
The default callback parses content into a data.frame while dropping other
response values to make the output easily passable to tidyverse packages like
dplyr or ggplot2. For further details or debugging one can pass a print or a
more compicated function.}

\item{model_id}{Character scalar formated as 'source-target-domain'.
Source language (Arabic, Brazilian Portuguese, English, French, Italian,
or Spanish), target language (Arabic, Brazilian Portuguese, English, French,
Italian, or Spanish) and domain of text (conversational, news, patent).
Check IBM documentation for other language mappings.}

\item{accept}{Character scalar that specifies response format. Alternative is
text/plain.}
}
\value{
Data.frame containing parsed content in a tidy fashion.
}
\description{
Translates text from Arabic, French,
  Portuguese or Spanish to English. Requires basic authentication using
  username and password.
}
\seealso{
Check \url{http://www.ibm.com/watson/developercloud/language-translation.html}
  for further documentation, and \url{https://language-translator-demo.mybluemix.net/}
  for a web demo.
}

